+++ tests/test_tap2sna.py
@@ -16,8 +16,9 @@ def mock_make_z80(*args):
     make_z80_args = args
 
 def mock_write_z80(ram, namespace, z80):
-    global snapshot
+    global snapshot, options
     snapshot = [0] * 16384 + ram
+    options = namespace
 
 class Tap2SnaTest(SkoolKitTestCase):
     def _write_tap(self, blocks, zip_archive=False, tap_name=None):
@@ -71,6 +72,7 @@ class Tap2SnaTest(SkoolKitTestCase):
         self.assertEqual([], options.ram_ops)
         self.assertEqual([], options.reg)
         self.assertIsNone(options.start)
+        self.assertFalse(options.sim_load)
         self.assertEqual([], options.state)
         self.assertEqual(options.user_agent, '')
 
@@ -706,6 +708,262 @@ class Tap2SnaTest(SkoolKitTestCase):
         output, error = self.run_tap2sna('--reg help')
         self.assertTrue(output.startswith('Usage: --reg name=value\n'))
         self.assertEqual(error, '')
+
+    @patch.object(tap2sna, '_write_z80', mock_write_z80)
+    def test_sim_load(self):
+        code_start = 32768
+        code_start_str = [ord(c) for c in str(code_start)]
+        basic_data = [
+            10, 0,            # Line 10
+            16, 0,            # Line length
+            239, 34, 34, 175, # LOAD ""CODE
+            58,               # :
+            249, 192, 176,    # RANDOMIZE USR VAL
+            34,               # "
+            *code_start_str,  # start address
+            34,               # "
+            13                # ENTER
+        ]
+        code = [4, 5]
+        blocks = [
+            create_tap_header_block("simloadbas", 10, len(basic_data), 0),
+            create_tap_data_block(basic_data),
+            create_tap_header_block("simloadbyt", code_start, len(code)),
+            create_tap_data_block(code)
+        ]
+        tapfile = self._write_tap(blocks)
+        z80file = 'out.z80'
+        output, error = self.run_tap2sna(f'--sim-load {tapfile} {z80file}')
+        out_lines = output.strip().split('\n')
+        exp_out_lines = [
+            'Program: simloadbas',
+            'Loading data block: 23755,20',
+            'Bytes: simloadbyt',
+            'Loading data block: 32768,2',
+            'Simulation ended: PC=32768',
+        ]
+        self.assertEqual(exp_out_lines, out_lines)
+        self.assertEqual(error, '')
+        self.assertEqual(basic_data, snapshot[23755:23755 + len(basic_data)])
+        self.assertEqual(code, snapshot[code_start:code_start + len(code)])
+        exp_reg = set(('SP=65344', 'IX=32770', 'IY=23610', 'PC=32768'))
+        self.assertTrue(exp_reg <= set(options.reg))
+
+    @patch.object(tap2sna, '_write_z80', mock_write_z80)
+    def test_sim_load_with_character_array(self):
+        code_start = 32768
+        code_start_str = [ord(c) for c in str(code_start)]
+        basic_data = [
+            10, 0,            # Line 10
+            25, 0,            # Line length
+            239, 34, 34, 228, # LOAD "" DATA
+            97, 36, 40, 41,   # a$()
+            58,               # :
+            239, 34, 34, 175, # LOAD ""CODE
+            58,               # :
+            249, 192, 176,    # RANDOMIZE USR VAL
+            34,               # "
+            *code_start_str,  # start address
+            34,               # "
+            13                # ENTER
+        ]
+        ca_name = "characters"
+        ca_data = [193, 5, 0, 1, 2, 0, 97, 98]
+        code = [4, 5]
+        blocks = [
+            create_tap_header_block("simloadbas", 10, len(basic_data), 0),
+            create_tap_data_block(basic_data),
+            create_tap_header_block(ca_name, length=len(ca_data), data_type=2),
+            create_tap_data_block(ca_data),
+            create_tap_header_block("simloadbyt", code_start, len(code)),
+            create_tap_data_block(code)
+        ]
+        tapfile = self._write_tap(blocks)
+        z80file = 'out.z80'
+        output, error = self.run_tap2sna(f'--sim-load {tapfile} {z80file}')
+        out_lines = output.strip().split('\n')
+        exp_out_lines = [
+            'Program: simloadbas',
+            'Loading data block: 23755,29',
+            'Character array: characters',
+            'Loading data block: 23787,8',
+            'Bytes: simloadbyt',
+            'Loading data block: 32768,2',
+            'Simulation ended: PC=32768',
+        ]
+        self.assertEqual(exp_out_lines, out_lines)
+        self.assertEqual(error, '')
+        self.assertEqual(basic_data, snapshot[23755:23755 + len(basic_data)])
+        self.assertEqual(ca_data, snapshot[23787:23787 + len(ca_data)])
+        self.assertEqual(code, snapshot[code_start:code_start + len(code)])
+        exp_reg = set(('SP=65344', 'IX=32770', 'IY=23610', 'PC=32768'))
+        self.assertTrue(exp_reg <= set(options.reg))
+
+    @patch.object(tap2sna, '_write_z80', mock_write_z80)
+    def test_sim_load_with_number_array(self):
+        code_start = 32768
+        code_start_str = [ord(c) for c in str(code_start)]
+        basic_data = [
+            10, 0,            # Line 10
+            24, 0,            # Line length
+            239, 34, 34, 228, # LOAD "" DATA
+            97, 40, 41,       # a()
+            58,               # :
+            239, 34, 34, 175, # LOAD ""CODE
+            58,               # :
+            249, 192, 176,    # RANDOMIZE USR VAL
+            34,               # "
+            *code_start_str,  # start address
+            34,               # "
+            13                # ENTER
+        ]
+        na_name = "numbers"
+        na_data = [129, 13, 0, 1, 2, 0, 0, 0, 1, 0, 0, 0, 0, 2, 0, 0]
+        code = [4, 5]
+        blocks = [
+            create_tap_header_block("simloadbas", 10, len(basic_data), 0),
+            create_tap_data_block(basic_data),
+            create_tap_header_block(na_name, length=len(na_data), data_type=1),
+            create_tap_data_block(na_data),
+            create_tap_header_block("simloadbyt", code_start, len(code)),
+            create_tap_data_block(code)
+        ]
+        tapfile = self._write_tap(blocks)
+        z80file = 'out.z80'
+        output, error = self.run_tap2sna(f'--sim-load {tapfile} {z80file}')
+        out_lines = output.strip().split('\n')
+        exp_out_lines = [
+            'Program: simloadbas',
+            'Loading data block: 23755,28',
+            'Number array: numbers   ',
+            'Loading data block: 23786,16',
+            'Bytes: simloadbyt',
+            'Loading data block: 32768,2',
+            'Simulation ended: PC=32768',
+        ]
+        self.assertEqual(exp_out_lines, out_lines)
+        self.assertEqual(error, '')
+        self.assertEqual(basic_data, snapshot[23755:23755 + len(basic_data)])
+        self.assertEqual(na_data, snapshot[23786:23786 + len(na_data)])
+        self.assertEqual(code, snapshot[code_start:code_start + len(code)])
+        exp_reg = set(('SP=65344', 'IX=32770', 'IY=23610', 'PC=32768'))
+        self.assertTrue(exp_reg <= set(options.reg))
+
+    @patch.object(tap2sna, '_write_z80', mock_write_z80)
+    def test_sim_load_with_headerless_block(self):
+        code_start = 32768
+        code_start_str = [ord(c) for c in str(code_start)]
+        basic_data = [
+            10, 0,            # Line 10
+            16, 0,            # Line length
+            239, 34, 34, 175, # LOAD ""CODE
+            58,               # :
+            249, 192, 176,    # RANDOMIZE USR VAL
+            34,               # "
+            *code_start_str,  # start address
+            34,               # "
+            13                # ENTER
+        ]
+        code = [
+            221, 33, 0, 192,  # LD IX,49152
+            17, 2, 0,         # LD DE,2
+            55,               # SCF
+            159,              # SBC A,A
+            221, 229,         # PUSH IX
+            195, 86, 5        # JP 1366
+        ]
+        code2 = [128, 129]
+        blocks = [
+            create_tap_header_block("simloadbas", 10, len(basic_data), 0),
+            create_tap_data_block(basic_data),
+            create_tap_header_block("simloadbyt", code_start, len(code)),
+            create_tap_data_block(code),
+            create_tap_data_block(code2)
+        ]
+        tapfile = self._write_tap(blocks)
+        z80file = 'out.z80'
+        output, error = self.run_tap2sna(f'--sim-load {tapfile} {z80file}')
+        out_lines = output.strip().split('\n')
+        exp_out_lines = [
+            'Program: simloadbas',
+            'Loading data block: 23755,20',
+            'Bytes: simloadbyt',
+            'Loading data block: 32768,14',
+            'Loading data block: 49152,2',
+            'Simulation ended: PC=49152',
+        ]
+        self.assertEqual(exp_out_lines, out_lines)
+        self.assertEqual(error, '')
+        self.assertEqual(basic_data, snapshot[23755:23755 + len(basic_data)])
+        self.assertEqual(code, snapshot[code_start:code_start + len(code)])
+        self.assertEqual(code2, snapshot[49152:49152 + len(code2)])
+        exp_reg = set(('SP=65344', 'IX=49154', 'IY=23610', 'PC=49152'))
+        self.assertTrue(exp_reg <= set(options.reg))
+
+    @patch.object(tap2sna, '_write_z80', mock_write_z80)
+    def test_sim_load_with_unexpected_block_type(self):
+        code_start = 32768
+        code_start_str = [ord(c) for c in str(code_start)]
+        basic_data = [
+            10, 0,            # Line 10
+            16, 0,            # Line length
+            239, 34, 34, 175, # LOAD ""CODE
+            58,               # :
+            249, 192, 176,    # RANDOMIZE USR VAL
+            34,               # "
+            *code_start_str,  # start address
+            34,               # "
+            13                # ENTER
+        ]
+        code = [4, 5]
+        blocks = [
+            create_tap_header_block("ublocktype", 10, len(basic_data), 0),
+            create_tap_data_block(basic_data),
+            create_tap_data_block(code)
+        ]
+        tapfile = self._write_tap(blocks)
+        z80file = 'out.z80'
+        with self.assertRaises(SkoolKitError) as cm:
+            self.run_tap2sna(f'--sim-load {tapfile} {z80file}')
+        self.assertEqual(cm.exception.args[0], f'Error while getting snapshot {z80file}: Failed to load block: unexpected type')
+
+    @patch.object(tap2sna, '_write_z80', mock_write_z80)
+    def test_sim_load_with_unexpected_data_length(self):
+        code_start = 32768
+        code_start_str = [ord(c) for c in str(code_start)]
+        basic_data = [
+            10, 0,            # Line 10
+            16,  0,           # Line length
+            239, 34, 34, 175, # LOAD ""CODE
+            58,               # :
+            249, 192, 176,    # RANDOMIZE USR VAL
+            34,               # "
+            *code_start_str,  # start address
+            34,               # "
+            13                # ENTER
+        ]
+        code = [4, 5]
+        c_len = len(code)
+        blocks = [
+            create_tap_header_block("badlength", 10, len(basic_data), 0),
+            create_tap_data_block(basic_data),
+            create_tap_header_block("badlength", code_start, c_len + 1),
+            create_tap_data_block(code)
+        ]
+        tapfile = self._write_tap(blocks)
+        z80file = 'out.z80'
+        with self.assertRaises(SkoolKitError) as cm:
+            self.run_tap2sna(f'--sim-load {tapfile} {z80file}')
+        self.assertEqual(cm.exception.args[0], f'Error while getting snapshot {z80file}: Failed to load block of length {c_len}: expected length {c_len + 1}')
+
+    @patch.object(tap2sna, '_write_z80', mock_write_z80)
+    def test_sim_load_with_unknown_block_type(self):
+        blocks = [create_tap_header_block("whatisthis", length=10, data_type=4)]
+        tapfile = self._write_tap(blocks)
+        z80file = 'out.z80'
+        with self.assertRaises(SkoolKitError) as cm:
+            self.run_tap2sna(f'--sim-load {tapfile} {z80file}')
+        self.assertEqual(cm.exception.args[0], f'Error while getting snapshot {z80file}: Failed to load block: unknown type: 4')
 
     def test_default_state(self):
         block = create_tap_data_block([0])